+++ rust/operator-binary/src/backend/k8s_search.rs
@@ -10,6 +10,7 @@ use stackable_operator::{
     },
     kube::api::ListParams,
     kvp::{LabelError, LabelSelectorExt, Labels},
+    utils::cluster_info::KubernetesClusterInfo,
 };
 
 use crate::{crd::SearchNamespace, format::SecretData, utils::Unloggable};
@@ -82,6 +83,7 @@ impl SecretBackend for K8sSearch {
 
     async fn get_secret_data(
         &self,
+        _cluster_info: &KubernetesClusterInfo,
         selector: &SecretVolumeSelector,
         pod_info: PodInfo,
     ) -> Result<SecretContents, Self::Error> {