+++ CHANGELOG.md
@@ -9,9 +9,8 @@ All notable changes to this project will be documented in this file.
 - Active Directory's `samAccountName` generation can now be customized ([#454]).
 - Added experimental cert-manager backend ([#482]).
 - Make RSA key length configurable ([#506]).
-- The operator can now run on Kubernetes clusters using a non-default cluster domain. It should automatically detect the
-  correct domain to use, but you can also use the env var `KUBERNETES_CLUSTER_DOMAIN` to set the domain explicitly
-  or use the helm-chart property `kubernetesClusterDomain` ([#510]).
+- The operator can now run on Kubernetes clusters using a non-default cluster domain.
+  Use the env var `KUBERNETES_CLUSTER_DOMAIN` or the operator Helm chart property `kubernetesClusterDomain` to set a non-default cluster domain` ([#510]).
 
 ### Changed